Refactor MediaEditorModal to improve image load error handling
- Introduced a new custom hook `useLoadErrorNotice` to manage image load error notifications and retries.
- Removed redundant state management related to image loading and error handling from `MediaEditorModalContent`.
- Updated the modal to utilize the new hook for clearer and more maintainable error handling logic.
- Enhanced user feedback by ensuring error notices are displayed appropriately during image loading failures.
