+++ packages/media-editor/src/components/media-editor-modal/index.tsx
@@ -53,6 +53,7 @@ import { getMediaTypeFromMimeType } from '../../utils';
 import { CropperProvider, useCropper } from '../../image-editor';
 import type { AspectRatioPreset } from '../../image-editor/core/constants';
 import { buildModifiers } from './build-modifiers';
+import { NOTICES_CONTEXT, useLoadErrorNotice } from './use-load-error-notice';
 
 // Details-tab edits the modal bundles into a transformed `/edit` request.
 // Matches Core's `WP_REST_Attachments_Controller::get_edit_media_item_args`
@@ -69,12 +70,7 @@ const METADATA_EDIT_KEYS = [
 	'post',
 ] as const;
 
-// Scope save-failure snackbars to this modal so they don't leak into the
-// host editor's notices tray (and vice versa).
-const NOTICES_CONTEXT = 'media-editor';
-const IMAGE_LOAD_ERROR_NOTICE_ID = 'media-editor-image-load-error';
 const PLACEMENT_CONTROL_IDLE_MS = 300;
-const LOAD_TIMEOUT_MS = 5000;
 
 const { Tabs } = unlock( componentsPrivateApis );
 
@@ -221,29 +217,21 @@ function MediaEditorModalContent( {
 	onUpdate,
 }: MediaEditorModalContentProps ) {
 	const cropper = useCropper();
-	const { setImage } = cropper;
 	const hasChanges = cropper.isDirty || hasEdits;
 
 	const registry = useRegistry();
 	const {
 		clearEntityRecordEdits,
 		editEntityRecord,
-		invalidateResolution,
 		receiveEntityRecords,
 		saveEditedEntityRecord,
 	} = useDispatch( coreStore );
 	const { closeMediaEditorModal } = useDispatch( mediaEditorStore );
-	const { createErrorNotice, removeAllNotices, removeNotice } =
-		useDispatch( noticesStore );
+	const { createErrorNotice, removeAllNotices } = useDispatch( noticesStore );
 
 	const [ isSaving, setIsSaving ] = useState( false );
 	const [ isDiscardDialogOpen, setIsDiscardDialogOpen ] = useState( false );
 	const [ isPlacementActive, setIsPlacementActive ] = useState( false );
-	const [ mediaLoadTimedOut, setMediaLoadTimedOut ] = useState( false );
-	const [ imageLoadAttempt, setImageLoadAttempt ] = useState( 0 );
-	const [ imageLoadState, setImageLoadState ] = useState<
-		'idle' | 'loading' | 'loaded' | 'error'
-	>( 'idle' );
 	const placementControlTimerRef =
 		useRef< ReturnType< typeof setTimeout > >();
 
@@ -274,14 +262,19 @@ function MediaEditorModalContent( {
 	const mediaType = getMediaTypeFromMimeType( media?.mime_type ).type;
 	const isImage = !! media && mediaType === 'image';
 	const mediaUrl = isImage ? media?.source_url : undefined;
-	const imageLoadFailed =
-		isImage && ( ! mediaUrl || imageLoadState === 'error' );
+	const canEditImage = isImage && !! mediaUrl;
 	const isImageLoading =
-		!! mediaUrl &&
-		! imageLoadFailed &&
-		( imageLoadState === 'loading' ||
-			cropper.state.image?.src !== mediaUrl );
-	const canEditImage = isImage && !! mediaUrl && ! imageLoadFailed;
+		canEditImage && cropper.state.image?.src !== mediaUrl;
+	const {
+		clearImageLoadErrorNotice,
+		imageLoadRetryKey,
+		showImageLoadErrorNotice,
+	} = useLoadErrorNotice( {
+		id,
+		hasMedia: !! media,
+		isImage,
+		mediaUrl,
+	} );
 
 	const imageAspectRatio = useMemo( () => {
 		if ( ! isImage ) {
@@ -299,174 +292,9 @@ function MediaEditorModalContent( {
 		return null;
 	}, [ isImage, media ] );
 
-	const clearImageLoadErrorNotice = useCallback( () => {
-		removeNotice( IMAGE_LOAD_ERROR_NOTICE_ID, NOTICES_CONTEXT );
-	}, [ removeNotice ] );
-
-	const showImageLoadErrorNotice = useCallback(
-		() =>
-			createErrorNotice(
-				__(
-					'Could not load image. Please check your connection and try again.'
-				),
-				{
-					id: IMAGE_LOAD_ERROR_NOTICE_ID,
-					type: 'snackbar',
-					context: NOTICES_CONTEXT,
-				}
-			),
-		[ createErrorNotice ]
-	);
-
-	const retryMediaLoad = useCallback( () => {
-		clearImageLoadErrorNotice();
-		setMediaLoadTimedOut( false );
-		invalidateResolution( 'getEntityRecord', [
-			'postType',
-			'attachment',
-			id,
-		] );
-		invalidateResolution( 'getEditedEntityRecord', [
-			'postType',
-			'attachment',
-			id,
-		] );
-	}, [ clearImageLoadErrorNotice, id, invalidateResolution ] );
-
-	const retryImageLoad = useCallback( () => {
-		clearImageLoadErrorNotice();
-		setImageLoadState( 'idle' );
-		setImageLoadAttempt( ( attempt ) => attempt + 1 );
-	}, [ clearImageLoadErrorNotice ] );
-
-	useEffect( () => {
-		retryMediaLoad();
-	}, [ retryMediaLoad ] );
-
-	useEffect( () => {
-		if ( media ) {
-			setMediaLoadTimedOut( false );
-			return;
-		}
-
-		const timeout = setTimeout(
-			() => {
-				setMediaLoadTimedOut( true );
-			},
-			navigator.onLine ? LOAD_TIMEOUT_MS : 0
-		);
-
-		return () => {
-			clearTimeout( timeout );
-		};
-	}, [ id, media ] );
-
-	useEffect( () => {
-		if ( mediaLoadTimedOut ) {
-			showImageLoadErrorNotice();
-		}
-	}, [ mediaLoadTimedOut, showImageLoadErrorNotice ] );
-
-	useEffect( () => {
-		if ( ! media ) {
-			setImageLoadState( 'idle' );
-			return;
-		}
-
-		if ( ! isImage ) {
-			setImageLoadState( 'idle' );
-			clearImageLoadErrorNotice();
-			return;
-		}
-
-		if ( ! mediaUrl ) {
-			setImageLoadState( 'error' );
-			showImageLoadErrorNotice();
-			return;
-		}
-
-		let isCurrent = true;
-		const image = new Image();
-
-		function handleLoad() {
-			if ( ! isCurrent ) {
-				return;
-			}
-			clearTimeout( timeout );
-			setImage( {
-				src: mediaUrl,
-				naturalWidth: image.naturalWidth,
-				naturalHeight: image.naturalHeight,
-			} );
-			setImageLoadState( 'loaded' );
-			clearImageLoadErrorNotice();
-		}
-
-		function handleError() {
-			if ( ! isCurrent ) {
-				return;
-			}
-			clearTimeout( timeout );
-			setImageLoadState( 'error' );
-			showImageLoadErrorNotice();
-		}
-
-		const timeout = setTimeout( handleError, LOAD_TIMEOUT_MS );
-		setImageLoadState( 'loading' );
-		image.addEventListener( 'load', handleLoad );
-		image.addEventListener( 'error', handleError );
-		image.src = mediaUrl;
-
-		if ( image.complete ) {
-			if ( image.naturalWidth > 0 ) {
-				handleLoad();
-			} else {
-				handleError();
-			}
-		}
-
-		return () => {
-			isCurrent = false;
-			clearTimeout( timeout );
-			image.removeEventListener( 'load', handleLoad );
-			image.removeEventListener( 'error', handleError );
-		};
-	}, [
-		clearImageLoadErrorNotice,
-		imageLoadAttempt,
-		isImage,
-		media,
-		mediaUrl,
-		setImage,
-		showImageLoadErrorNotice,
-	] );
-
 	const handleImageLoaded = useCallback( () => {
-		setImageLoadState( 'loaded' );
 		clearImageLoadErrorNotice();
 	}, [ clearImageLoadErrorNotice ] );
-
-	const handleImageLoadError = useCallback( () => {
-		setImageLoadState( 'error' );
-		showImageLoadErrorNotice();
-	}, [ showImageLoadErrorNotice ] );
-
-	useEffect( () => {
-		const handleOnline = () => {
-			if ( ! media ) {
-				retryMediaLoad();
-				return;
-			}
-			if ( isImage ) {
-				retryImageLoad();
-			}
-		};
-
-		window.addEventListener( 'online', handleOnline );
-		return () => {
-			window.removeEventListener( 'online', handleOnline );
-		};
-	}, [ isImage, media, retryImageLoad, retryMediaLoad ] );
 
 	const tabs = useMemo< ModalTab[] >( () => {
 		const detailsTab: ModalTab = {
@@ -647,6 +475,7 @@ function MediaEditorModalContent( {
 		canvasContent = (
 			<>
 				<MediaEditorCanvas
+					key={ imageLoadRetryKey }
 					aspectRatio={ resolveAspectRatio(
 						aspectRatioValue,
 						imageAspectRatio
@@ -654,7 +483,7 @@ function MediaEditorModalContent( {
 					freeformCrop={ freeformCrop }
 					isPlacementActive={ isPlacementActive }
 					onImageLoaded={ handleImageLoaded }
-					onImageLoadError={ handleImageLoadError }
+					onImageLoadError={ showImageLoadErrorNotice }
 				/>
 				{ isImageLoading && (
 					<div className="media-editor-modal__loading media-editor-modal__loading--overlay">
